backend/requirements.txt
alabaster==0.7.13
asgiref==3.6.0
Babel==2.12.1
bootstrap-py==2.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.1
django-cors-headers==3.14.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
docutils==0.20.1
drf-yasg==1.21.5
filelock==3.12.0
gitdb==4.0.10
GitPython==3.1.31
gunicorn==20.1.0
idna==3.4
imagesize==1.4.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
pguard==0.4.0
platformdirs==3.5.1
psycopg2-binary==2.9.6
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
setuptools-scm==7.1.0
smmap==5.0.0
snowballstemmer==2.2.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
Sphinx==7.0.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.23.0
